refactor(browser): split BrowserCookieStore into cookie_store.py (S6)
